hybrid_store/events/requirements.txt
kafka-python==2.0.2
pandas
pyarrow